requirements.txt
aiogram==3.13.1
anthropic==0.40.0
requests==2.32.3
beautifulsoup4==4.12.3
apscheduler==3.10.4
python-dotenv==1.0.1
lxml==5.3.0
playwright==1.49.0
telethon==1.37.0
